+++ server/wordi/src/main/java/com/pm/wordi/controller/dto/MentoringDto.java
@@ -2,24 +2,21 @@ package com.pm.wordi.controller.dto;
 
 import com.pm.wordi.controller.dto.MentorDto.ScheduleDTO;
 import com.pm.wordi.domain.BaseStatus;
-import com.pm.wordi.domain.mentor.Mentor;
-import com.pm.wordi.domain.mentor.MentorSchedule;
-import com.pm.wordi.domain.mentoring.Mentoring;
-import com.pm.wordi.domain.mentoring.MentoringStatus;
-import com.pm.wordi.domain.mentoring.Payment;
-import com.pm.wordi.domain.mentoring.PaymentStatus;
-import com.pm.wordi.domain.user.User;
+import com.pm.wordi.domain.mentor.entity.Mentor;
+import com.pm.wordi.domain.mentoring.entity.Mentoring;
+import com.pm.wordi.domain.mentoring.entity.MentoringStatus;
+import com.pm.wordi.domain.mentoring.entity.Payment;
+import com.pm.wordi.domain.mentoring.entity.PaymentStatus;
+import com.pm.wordi.domain.user.entity.User;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.lang.Nullable;
-import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
