+++ server/wordi/src/main/java/com/pm/wordi/domain/mentor/entity/MentorLevel.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.domain.mentor;
+package com.pm.wordi.domain.mentor.entity;
 
 public enum MentorLevel {
     대기, 일반, 인기
