+++ server/wordi/src/main/java/com/pm/wordi/domain/mentor/entity/MentorSchedule.java
@@ -1,10 +1,7 @@
-package com.pm.wordi.domain.review;
+package com.pm.wordi.domain.mentor.entity;
 
-import com.pm.wordi.domain.BaseStatus;
-import com.pm.wordi.domain.BaseTimeEntity;
-import com.pm.wordi.domain.mentor.Mentor;
-import com.pm.wordi.domain.user.User;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -13,29 +10,29 @@ import javax.persistence.*;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Review extends BaseTimeEntity {
+public class MentorSchedule {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "reviewId")
+    @Column(name = "mentorScheduleId")
     private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "userId")
-    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mentorId")
     private Mentor mentor;
 
-    private String content;
+    private String week;
 
-    @Enumerated(EnumType.STRING)
-    private BaseStatus status;
+    private String schedule;
 
     public void updateMentor(Mentor mentor) {
         this.mentor = mentor;
     }
 
-
+    @Builder
+    public MentorSchedule(Mentor mentor, String week, String schedule) {
+        this.mentor = mentor;
+        this.week = week;
+        this.schedule = schedule;
+    }
 }
