+++ server/wordi/src/main/java/com/pm/wordi/domain/mentor/repository/MentorKeywordRepository.java
@@ -1,5 +1,7 @@
-package com.pm.wordi.domain.mentor;
+package com.pm.wordi.domain.mentor.repository;
 
+import com.pm.wordi.domain.mentor.entity.Mentor;
+import com.pm.wordi.domain.mentor.entity.MentorKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MentorKeywordRepository extends JpaRepository<MentorKeyword, Long> {
