+++ server/wordi/src/main/java/com/pm/wordi/domain/mentoring/entity/MentoringStatus.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.domain.mentoring;
+package com.pm.wordi.domain.mentoring.entity;
 
 public enum MentoringStatus {
     승인대기, 예약확정, 예약거절, 멘토링완료, 예약취소
