+++ server/wordi/src/main/java/com/pm/wordi/domain/mentoring/entity/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.domain.mentoring;
+package com.pm.wordi.domain.mentoring.entity;
 
 public enum PaymentStatus {
     결제보류, 결제완료, 결제취소, 환불
