+++ server/wordi/src/main/java/com/pm/wordi/domain/mentoring/repository/PaymentRepository.java
@@ -1,5 +1,6 @@
-package com.pm.wordi.domain.mentoring;
+package com.pm.wordi.domain.mentoring.repository;
 
+import com.pm.wordi.domain.mentoring.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
