+++ server/wordi/src/main/java/com/pm/wordi/domain/user/entity/UserKeyword.java
@@ -1,6 +1,5 @@
-package com.pm.wordi.domain.user;
+package com.pm.wordi.domain.user.entity;
 
-import com.pm.wordi.domain.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
