+++ server/wordi/src/main/java/com/pm/wordi/domain/user/entity/UserLevel.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.domain.user;
+package com.pm.wordi.domain.user.entity;
 
 public enum UserLevel {
     관리자, 일반
