+++ server/wordi/src/main/java/com/pm/wordi/exception/certification/DecryptException.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.exception;
+package com.pm.wordi.exception.certification;
 
 public class DecryptException extends Exception{
     public DecryptException() {
