+++ server/wordi/src/main/java/com/pm/wordi/exception/certification/EncryptException.java
@@ -1,4 +1,4 @@
-package com.pm.wordi.exception;
+package com.pm.wordi.exception.certification;
 
 public class EncryptException extends Exception{
     public EncryptException() {
